feat: display products list

+++ back-end/src/database/migrations/20220804192114-create-sale.js
@@ -28,7 +28,7 @@ module.exports = {
         }
       },
       total_price: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(9, 2)
       },
       delivery_address: {
         type: Sequelize.STRING
