+++ back-end/src/database/migrations/20220804192248-create-product.js
@@ -12,7 +12,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       price: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(4, 2)
       },
       url_image: {
         type: Sequelize.STRING
